update EditData
i will add css
