+++ Traviary/src/components/Contents.tsx
@@ -1,18 +1,11 @@
 /** @format */
-import React, {
-	useEffect,
-	useState,
-	useRef,
-	ChangeEvent,
-	FormEvent,
-} from "react"
+import React, { useEffect, useState } from "react"
+import EditData from "./EditData"
 import styled, { keyframes, css } from "styled-components"
 import { doc, deleteDoc, updateDoc } from "firebase/firestore"
 import { AiTwotoneDelete } from "react-icons/ai"
 import { BiPencil } from "react-icons/bi"
-import { dbService, storageService } from "../fbase"
-import { deleteObject, ref } from "firebase/storage"
-import { ImStarFull } from "react-icons/im"
+import { dbService } from "../fbase"
 import { TbNotes } from "react-icons/tb"
 import { BsStars } from "react-icons/bs"
 import ReadStar from "./ReadStar"
@@ -31,43 +24,18 @@ const Contents = ({
 	const [open, setOpen] = useState(false)
 	const [editing, setEditing] = useState(false)
 	const [editData, setEditData] = useState(false)
-	const [editText, setEditText] = useState(traviObj.text)
-	const [editFile, setEditFile] = useState(traviObj.fileAttachURL)
 
 	const TraviRef = doc(dbService, "TraviDB", `${traviObj.id}`)
-	const urlRef = ref(storageService, traviObj.fileAttachURL)
 
 	const onDeleteClick = async () => {
 		const Ok = window.confirm("삭제하시겠습니까???")
 		if (Ok) {
 			await deleteDoc(TraviRef)
-			await deleteObject(urlRef)
 		}
 	}
 
 	const toggleEditing = () => {
 		setEditData((prev) => !prev)
-	}
-
-	const onChangeText = (
-		event: ChangeEvent<HTMLInputElement> | ChangeEvent<HTMLTextAreaElement>
-	) => {
-		const {
-			target: { value },
-		} = event
-		setEditText(value)
-	}
-	const onChangeFile = (event: ChangeEvent<HTMLInputElement>) => {
-		const {
-			currentTarget: { files },
-		} = event as any
-		setEditFile(files)
-	}
-
-	const onSubmit = async (event: FormEvent) => {
-		event.preventDefault()
-		await updateDoc(TraviRef, { text: editText, fileAttachUrl: editFile })
-		setEditData(false)
 	}
 
 	const handleClose = () => {
@@ -109,21 +77,7 @@ const Contents = ({
 										/>
 										{editData ? (
 											<>
-												<form onSubmit={onSubmit}>
-													<input
-														type="textarea"
-														value={traviObj.text}
-														onChange={onChangeText}
-														required
-													/>
-													<input
-														type="file"
-														accept="image/*"
-														onChange={onChangeFile}
-														required
-													/>
-													<input type="submit" value="update" />
-												</form>
+												<EditData traviObj={traviObj} />
 											</>
 										) : (
 											<></>
